% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{as.data.frame.CMOCounterfactual}
\alias{as.data.frame.CMOCounterfactual}
\title{The counterfactual contrasts, as a data frame}
\usage{
\method{as.data.frame}{CMOCounterfactual}(x, ...)
}
\arguments{
\item{x}{A \code{CMOCounterfactual} object.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame}, one row per modifiable variable.
}
\description{
The counterfactual contrasts, as a data frame
}
\seealso{
\code{\link{counterfactual}}
}
