% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{BlockDiagnostics}
\alias{BlockDiagnostics}
\title{Create a BlockDiagnostics object}
\usage{
BlockDiagnostics()
}
\value{
A BlockDiagnostics object.
}
\description{
Stores every statistic calculated for one data block.
}
\keyword{internal}
