% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{CMOValidation}
\alias{CMOValidation}
\title{Create a CMOValidation object}
\usage{
CMOValidation()
}
\value{
A CMOValidation object.
}
\description{
Stores every result generated during data validation.
}
\keyword{internal}
