% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.check_plan_matches}
\alias{.check_plan_matches}
\title{Refuse to execute a plan that does not describe this object}
\usage{
.check_plan_matches(object, plan, recipes, force)
}
\description{
Refuse to execute a plan that does not describe this object
}
\keyword{internal}
