% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.detect_variable_type}
\alias{.detect_variable_type}
\title{Detect the statistical nature of a single variable}
\usage{
.detect_variable_type(x)
}
\description{
Detect the statistical nature of a single variable
}
\keyword{internal}
