% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.html_table}
\alias{.html_table}
\title{Render a data.frame as an interactive HTML table}
\usage{
.html_table(df, caption = NULL)
}
\description{
Render a data.frame as an interactive HTML table
}
\keyword{internal}
