% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.identification_assumptions}
\alias{.identification_assumptions}
\title{Assumptions implied by an identification strategy}
\usage{
.identification_assumptions(identification, adjustment_set = character())
}
\description{
Assumptions implied by an identification strategy
}
\keyword{internal}
