% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.interpret_graph}
\alias{.interpret_graph}
\title{Extract readable findings from the graph}
\usage{
.interpret_graph(graph, importance, outcome_name, params)
}
\description{
Extract readable findings from the graph
}
\keyword{internal}
