% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.preprocess_block}
\alias{.preprocess_block}
\title{Run the full recipe for a single block}
\usage{
.preprocess_block(block_name, x, recipe, context, index_offset = 0L)
}
\description{
Run the full recipe for a single block
}
\keyword{internal}
