% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.rank_reference}
\alias{.rank_reference}
\title{Reference distribution used to map new values onto training ranks}
\usage{
.rank_reference(x)
}
\description{
Reference distribution used to map new values onto training ranks
}
\keyword{internal}
