% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.report_bullets}
\alias{.report_bullets}
\title{Print a character vector as a bulleted, wrapped list}
\usage{
.report_bullets(x, prefix = "  - ", width = 78, max_items = Inf)
}
\description{
Print a character vector as a bulleted, wrapped list
}
\keyword{internal}
