% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.resolve_outcome}
\alias{.resolve_outcome}
\title{Resolve the outcome from metadata}
\usage{
.resolve_outcome(metadata, outcome, sample_ids)
}
\description{
Resolve the outcome from metadata
}
\keyword{internal}
