% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.stage_config}
\alias{.stage_config}
\title{Method configured for a stage in a recipe}
\usage{
.stage_config(recipe, stage)
}
\description{
Method configured for a stage in a recipe
}
\keyword{internal}
