% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.structural_sample_filter}
\alias{.structural_sample_filter}
\title{Decide which samples the structural filters remove}
\usage{
.structural_sample_filter(x, recipe)
}
\description{
Decide which samples the structural filters remove
}
\keyword{internal}
