% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MultiOmicsData}
\alias{MultiOmicsData}
\title{Create a MultiOmicsData object}
\usage{
MultiOmicsData()
}
\value{
A MultiOmicsData object.
}
\description{
Core container storing every data block loaded into CausalMultiOmics.
}
\keyword{internal}
