% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{PreprocessingRecipe}
\alias{PreprocessingRecipe}
\title{Create a PreprocessingRecipe object}
\usage{
PreprocessingRecipe()
}
\value{
A PreprocessingRecipe object.
}
\description{
Stores the preprocessing strategy recommended during validation or
specified manually by the user.
}
\keyword{internal}
