% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{PreprocessingResult}
\alias{PreprocessingResult}
\title{Create a PreprocessingResult object}
\usage{
PreprocessingResult()
}
\value{
A PreprocessingResult object.
}
\description{
Stores every output generated during preprocessing.
}
\keyword{internal}
