% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{print.analysis_assumptions}
\alias{print.analysis_assumptions}
\title{Print what is being assumed}
\usage{
\method{print}{analysis_assumptions}(x, ...)
}
\arguments{
\item{x}{An \code{analysis_assumptions}.}

\item{...}{Ignored.}
}
\value{
The object, invisibly.
}
\description{
Print what is being assumed
}
