% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{print.analysis_control}
\alias{print.analysis_control}
\title{Print how much work will be done}
\usage{
\method{print}{analysis_control}(x, ...)
}
\arguments{
\item{x}{An \code{analysis_control}.}

\item{...}{Ignored.}
}
\value{
The object, invisibly.
}
\description{
Print how much work will be done
}
