% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{print.CMOComparison}
\alias{print.CMOComparison}
\title{Print a comparison of two analyses}
\usage{
\method{print}{CMOComparison}(x, ...)
}
\arguments{
\item{x}{A \code{CMOComparison}.}

\item{...}{Ignored.}
}
\value{
The object, invisibly.
}
\description{
Print a comparison of two analyses
}
