% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{print.CMOMultiResult}
\alias{print.CMOMultiResult}
\title{Print an analysis of several outcomes}
\usage{
\method{print}{CMOMultiResult}(x, ...)
}
\arguments{
\item{x}{A \code{CMOMultiResult}.}

\item{...}{Ignored.}
}
\value{
The object, invisibly.
}
\description{
Print an analysis of several outcomes
}
