% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ModuleGraph}
\alias{print.ModuleGraph}
\title{Print the groups of features that move together}
\usage{
\method{print}{ModuleGraph}(x, ...)
}
\arguments{
\item{x}{A \code{ModuleGraph}.}

\item{...}{Ignored.}
}
\value{
The object, invisibly.
}
\description{
Print the groups of features that move together
}
