% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TransformationRecommendation}
\alias{TransformationRecommendation}
\title{Create a TransformationRecommendation object}
\usage{
TransformationRecommendation()
}
\value{
A TransformationRecommendation object.
}
\description{
Stores automatic transformation benchmarking for one data block.
}
\keyword{internal}
